requirements.txt
streamlit>=1.29.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
transformers>=4.36.0
torch>=2.1.0
python-multipart>=0.0.6
aiofiles>=23.2.1
scipy>=1.11.0
